Extract I18NConfigurator from I18NAcceptLanguageFilter
(ServletContext injection for JAX-RS provider does not work on Quarkus)
